tools/bazel.rc: Moved .bazelrc to tools subdir.
This is a 'global' file specific to the workspace. The change is just in which order the file will be read; being in workspace root means it's read after system rc, and before homedir rc and user-specific rc (passed by a flag), while being in workspace/tools/bazel.rc means it's read after etc/bazel.bazelrc, but before homedir one. Presumably the 'global' versions are read before 'local' versions.